# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gyroscope real-time monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /serial/device-info:
    post:
      operationId: postDeviceInfo
      tags: [health]
      summary: Register parsed DEVICE_INFO / HEALTH handshake from firmware
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInfoInput"
      responses:
        "200":
          description: Capabilities derived from device info
          content:
            application/json:
              schema:
                type: object
                required: [ok, capabilities]
                properties:
                  ok:
                    type: boolean
                  capabilities:
                    $ref: "#/components/schemas/Capabilities"
...
components:
  schemas:
...
    Capabilities:
      type: object
      required: [imu, motor, temperature, vibration, sdcard]
      properties:
        imu:
          type: boolean
          description: MPU6050 / IMU / tilt sensor present
        motor:
          type: boolean
          description: ESC / motor controller present
        temperature:
          type: boolean
          description: Temperature sensor present
        vibration:
          type: boolean
          description: Vibration sensor present
        sdcard:
          type: boolean
          description: SD card module present
    DeviceInfoInput:
      type: object
      properties:
        deviceId:
          type: string
        version:
          type: string
        baud:
          type: integer
        components:
          type: array
          items:
            type: string
        health:
          type: object
          additionalProperties:
            type: string
            enum: [OK, FAIL, UNKNOWN]
...